PHOENIX-4683 Cap timeouts for stats precompact hook logic
